Added phrases
Added some more phrases for taking a piece and moving out of check.
Replaced RESPONSE NOT FORMULATED with human responses, in case a correct
move is not found as this looks much more human should it occur. Also
changed typing time as it seemed much too short considering the chess
tutor should have to think of the next move, as well as then having to
type it.