Minor change
Added a quick end game status message, displays where the responses show
up.
This occurs when the game ends in either a checkmate or draw, and will
output the directly correlating 'status' variable.

This may change perhaps to grant the user a little advice on how/why the
game has ended