Added More End Game Responses
More tutor responses have been added for the end-game states of player
victory, loss and draws. (Currently 3 responses for each state)
These messages are selected based on a random number generator at the
moment.

Further Implementation: Need to account for the number of turns it has
taken for the game to end